// types/admin-settings.ts
export type HeroSlide = {
    title: string;
    subtitle: string;
    image: string;
    price: string;
};

export type CollectionGridItem = {
    image: string;
    link: string;
};

export type AdminSettingsData = {
    id: string;
    storeName: string;
    storeEmail: string | null;
    storePhone: string | null;
    storeAddress: string | null;
    metaTitle: string | null;
    metaDescription: string | null;
    facebookUrl: string | null;
    instagramUrl: string | null;
    youtubeUrl: string | null;
    lowStockThreshold: number;
    stockReservationMinutes: number;
    logoUrl: string | null;
    faviconUrl: string | null;
    updatedAt: string;
    heroSlides: HeroSlide[];
    heroCtaText: string | null;
    heroCtaLink: string | null;
    headerAnnouncementEnabled: boolean;
    headerAnnouncementText: string | null;
    headerAnnouncementCtaText: string | null;
    headerAnnouncementCtaLink: string | null;
    headerBusinessHours: string | null;
    footerTagline: string | null;
    footerCopyright: string | null;
    footerTwitterUrl: string | null;
    footerLinkedinUrl: string | null;
    productsPerPage: number;
    showOutOfStock: boolean;
    homepageTrendingTitle: string | null;
    homepageTrendingSubtitle: string | null;
    homepageRecentTitle: string | null;
    homepageRecentSubtitle: string | null;
    homepageFestiveEyebrow: string | null;
    homepageFestiveTitle: string | null;
    homepageFestiveCtaText: string | null;
    homepageFestiveCtaLink: string | null;
    homepageFestiveImage: string | null;
    homepageLatestNewsTitle: string | null;
    homepageLatestNewsSubtitle: string | null;
    homepageFeaturedProductId: string | null;
    homepageProductCardIds: string[];
    homepageTrendingProductIds: string[];
    homepageRecentProductIds: string[];
    homepageCollectionImages: CollectionGridItem[];
};

export type AdminSettingsFormData = {
    storeName: string;
    storeEmail: string;
    storePhone: string;
    storeAddress: string;
    metaTitle: string;
    metaDescription: string;
    facebookUrl: string;
    instagramUrl: string;
    youtubeUrl: string;
    lowStockThreshold: number;
    stockReservationMinutes: number;
    logoUrl: string;
    faviconUrl: string;
    heroTitle1: string;
    heroSubtitle1: string;
    heroImage1: string;
    heroPrice1: string;
    heroTitle2: string;
    heroSubtitle2: string;
    heroImage2: string;
    heroPrice2: string;
    heroTitle3: string;
    heroSubtitle3: string;
    heroImage3: string;
    heroPrice3: string;
    heroCtaText: string;
    heroCtaLink: string;
    headerAnnouncementEnabled: boolean;
    headerAnnouncementText: string;
    headerAnnouncementCtaText: string;
    headerAnnouncementCtaLink: string;
    headerBusinessHours: string;
    footerTagline: string;
    footerCopyright: string;
    footerTwitterUrl: string;
    footerLinkedinUrl: string;
    productsPerPage: number;
    showOutOfStock: boolean;
    homepageTrendingTitle: string;
    homepageTrendingSubtitle: string;
    homepageRecentTitle: string;
    homepageRecentSubtitle: string;
    homepageFestiveEyebrow: string;
    homepageFestiveTitle: string;
    homepageFestiveCtaText: string;
    homepageFestiveCtaLink: string;
    homepageFestiveImage: string;
    homepageLatestNewsTitle: string;
    homepageLatestNewsSubtitle: string;
    homepageFeaturedProductId: string;
    homepageProductCardIds: string;
    homepageTrendingProductIds: string;
    homepageRecentProductIds: string;
    homepageCollectionImage1: string;
    homepageCollectionLink1: string;
    homepageCollectionImage2: string;
    homepageCollectionLink2: string;
    homepageCollectionImage3: string;
    homepageCollectionLink3: string;
    homepageCollectionImage4: string;
    homepageCollectionLink4: string;
    homepageCollectionImage5: string;
    homepageCollectionLink5: string;
};